+++ NEWS.md
@@ -1,3 +1,5 @@
+# ragg (development version)
+
 # ragg 1.3.0
 
 * Added supported for new graphics enigine features: Groups, paths, luminance