+++ NEWS.md
@@ -3,6 +3,7 @@
 * Make sure the linejoin/linemitre setting is honored when drawing rectangles
   (#162)
 * Fix a bug with resetting clipping during group and pattern rendering
+* Fix scaling of raster glyphs in drawGlyph
 
 # ragg 1.3.0