+++ NEWS.md
@@ -1,4 +1,4 @@
-# ragg (development version)
+# ragg 1.3.1
 
 * Make sure the linejoin/linemitre setting is honored when drawing rectangles
   (#162)